requirements.txt
Flask==2.1.1
#psycopg2==2.9.3
psycopg2-binary==2.9.3
SQLAlchemy==1.4.28
Flask-SQLAlchemy==2.5.1
